+++ MozuAndroidInStoreAssistant/app/src/main/java/com/mozu/mozuandroidinstoreassistant/app/data/product/FulfillmentInfo.java
@@ -3,12 +3,13 @@ package com.mozu.mozuandroidinstoreassistant.app.data.product;
 public class FulfillmentInfo {
 
 
-    public FulfillmentInfo(String type,String location){
+    public String mType;
+    public String mLocation;
+
+    public FulfillmentInfo(String type, String location) {
         mType = type;
         mLocation = location;
     }
-    public String mType;
-    public String mLocation;
 
 
 }
