+++ MozuAndroidInStoreAssistant/app/src/main/java/com/mozu/mozuandroidinstoreassistant/app/order/NewOrderCreateFragment.java
@@ -33,6 +33,7 @@ import com.mozu.mozuandroidinstoreassistant.app.data.order.OrderItemRow;
 import com.mozu.mozuandroidinstoreassistant.app.layout.order.NewOrderShippingItemLayout;
 import com.mozu.mozuandroidinstoreassistant.app.models.authentication.UserAuthenticationStateMachine;
 import com.mozu.mozuandroidinstoreassistant.app.models.authentication.UserAuthenticationStateMachineProducer;
+import com.mozu.mozuandroidinstoreassistant.app.order.adapters.NewOrderProductAdapter;
 import com.mozu.mozuandroidinstoreassistant.app.order.adapters.ProductSuggestionAdapter;
 import com.mozu.mozuandroidinstoreassistant.app.order.loaders.NewOrderManager;
 
@@ -49,22 +50,20 @@ import rx.schedulers.Schedulers;
 
 public class NewOrderCreateFragment extends Fragment implements NewOrderItemEditFragment.onItemEditDoneListener, NewOrderShippingItemLayout.OrderUpdateListener {
 
-    private View mView;
+    private static final String EDIT_MODE = "editMode";
     @InjectView(R.id.product_lookup)
     public AutoCompleteTextView mProductLookup;
 
     @InjectView(R.id.product_listview)
     public ListView mOrderProducts;
-
-    private int mTenantId;
-    private int mSiteId;
     Subscription mSubscription;
     ProductSuggestionAdapter mAdapter;
-    private com.mozu.mozuandroidinstoreassistant.app.order.NewOrderProductAdapter mProductsAdapter;
+    private View mView;
+    private int mTenantId;
+    private int mSiteId;
+    private NewOrderProductAdapter mProductsAdapter;
     private Order mOrder;
     private boolean mIsEditMode;
-    private static final String EDIT_MODE = "editMode";
-
 
     public static NewOrderCreateFragment getInstance(boolean isEditMode) {
         NewOrderCreateFragment newOrderCreateFragment = new NewOrderCreateFragment();
@@ -72,165 +71,6 @@ public class NewOrderCreateFragment extends Fragment implements NewOrderItemEdit
         b.putBoolean(EDIT_MODE, isEditMode);
         newOrderCreateFragment.setArguments(b);
         return newOrderCreateFragment;
-    }
-
-    public void setOrder(Order order) {
-        mOrder = order;
-    }
-
-
-    @Override
-    public void setUserVisibleHint(final boolean isVisibleToUser) {
-        super.setUserVisibleHint(isVisibleToUser);
-        new Handler().post(new Runnable() {
-            @Override
-            public void run() {
-                if (getActivity() instanceof OrderDetailActivity) {
-                    ((OrderDetailActivity) getActivity()).setEditModeVisibility(isVisibleToUser);
-                }
-            }
-        });
-    }
-
-    @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        setUpViews();
-    }
-
-    @Override
-    public void onSaveInstanceState(Bundle outState) {
-        outState.putBoolean("hsjdhsj", true);
-        super.onSaveInstanceState(outState);
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setRetainInstance(true);
-        RxBus.getInstance().toObserverable()
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(getEventSubscriber());
-    }
-
-    @Nullable
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        mView = inflater.inflate(R.layout.new_order_details_fragment, null);
-        UserAuthenticationStateMachine mUserState = UserAuthenticationStateMachineProducer.getInstance(getActivity());
-        mSiteId = mUserState.getSiteId();
-        mTenantId = mUserState.getTenantId();
-        mIsEditMode = getArguments().getBoolean(EDIT_MODE);
-        ButterKnife.inject(this, mView);
-        NewOrderManager.count = 0;
-        return mView;
-    }
-
-    private Subscriber<Object> getEventSubscriber() {
-        return new Subscriber<Object>() {
-            @Override
-            public void onCompleted() {
-
-            }
-
-            @Override
-            public void onError(Throwable e) {
-
-            }
-
-            @Override
-            public void onNext(Object o) {
-                if (o instanceof Boolean) {
-                    mIsEditMode = (Boolean) o;
-                    updateEditMode(mIsEditMode);
-                }
-            }
-        };
-    }
-
-    private void updateEditMode(Boolean editMode) {
-        mProductsAdapter.setEditMode(editMode);
-        mProductsAdapter.notifyDataSetChanged();
-        mProductLookup.setVisibility(editMode ? View.VISIBLE : View.GONE);
-    }
-
-
-    private void setUpViews() {
-        mAdapter = new ProductSuggestionAdapter(getActivity());
-        mProductLookup.setAdapter(mAdapter);
-        mProductsAdapter = new com.mozu.mozuandroidinstoreassistant.app.order.NewOrderProductAdapter(this);
-        if (mOrder != null && mOrder.getItems() != null) {
-            mProductsAdapter.addData(mOrder);
-            updateEditMode(mIsEditMode);
-        }
-
-
-        mOrderProducts.setAdapter(mProductsAdapter);
-        mProductLookup.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence charSequence, int i, int i2, int i3) {
-
-            }
-
-            @Override
-            public void onTextChanged(final CharSequence charSequence, int start, int before, int count) {
-                if (mSubscription != null && !mSubscription.isUnsubscribed()) {
-                    mSubscription.unsubscribe();
-                }
-
-                mSubscription = AndroidObservable.bindFragment(NewOrderCreateFragment.this, NewOrderManager.getInstance().getProductSuggestion(charSequence.toString(), mTenantId, mSiteId))
-                        .subscribe(new Subscriber<ProductSearchResult>() {
-                            @Override
-                            public void onCompleted() {
-
-                            }
-
-                            @Override
-                            public void onError(Throwable e) {
-                                Toast.makeText(getActivity(), "Error Searching for product " + charSequence, Toast.LENGTH_SHORT).show();
-                            }
-
-                            @Override
-                            public void onNext(ProductSearchResult productSearchResult) {
-                                List<Product> data = productSearchResult.getItems();
-                                mAdapter.clear();
-                                mAdapter.addAll(data);
-                                mAdapter.notifyDataSetChanged();
-                            }
-                        });
-            }
-
-            @Override
-            public void afterTextChanged(Editable editable) {
-
-            }
-        });
-
-        mProductLookup.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
-                Product product = (Product) adapterView.getItemAtPosition(position);
-                OrderItem orderItem = new OrderItem();
-                orderItem.setProduct(convertProduct(product));
-                orderItem.setQuantity(1);
-                NewOrderItemEditFragment newOrderItemEditFragment = NewOrderItemEditFragment.getInstance(orderItem, mOrder.getId(), false);
-                newOrderItemEditFragment.setOnEditDoneListener(NewOrderCreateFragment.this);
-                newOrderItemEditFragment.show(getFragmentManager(), "");
-            }
-        });
-
-        mOrderProducts.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
-                if (mIsEditMode && adapterView.getItemAtPosition(position) instanceof OrderItemRow) {
-                    OrderItem orderItem = ((OrderItemRow) adapterView.getItemAtPosition(position)).orderItem;
-                    NewOrderItemEditFragment newOrderItemEditFragment = NewOrderItemEditFragment.getInstance(orderItem, mOrder.getId(), true);
-                    newOrderItemEditFragment.setOnEditDoneListener(NewOrderCreateFragment.this);
-                    newOrderItemEditFragment.show(getFragmentManager(), "");
-                }
-            }
-        });
     }
 
     public static com.mozu.api.contracts.commerceruntime.products.Product convertProduct(Product inProduct) {
@@ -390,6 +230,163 @@ public class NewOrderCreateFragment extends Fragment implements NewOrderItemEdit
             pkgMeasurements.setWidth(measurements.getPackageWidth());
         }
         return pkgMeasurements;
+    }
+
+    public void setOrder(Order order) {
+        mOrder = order;
+    }
+
+    @Override
+    public void setUserVisibleHint(final boolean isVisibleToUser) {
+        super.setUserVisibleHint(isVisibleToUser);
+        new Handler().post(new Runnable() {
+            @Override
+            public void run() {
+                if (getActivity() instanceof OrderDetailActivity) {
+                    ((OrderDetailActivity) getActivity()).setEditModeVisibility(isVisibleToUser);
+                }
+            }
+        });
+    }
+
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        setUpViews();
+    }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        outState.putBoolean("hsjdhsj", true);
+        super.onSaveInstanceState(outState);
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setRetainInstance(true);
+        RxBus.getInstance().toObserverable()
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(getEventSubscriber());
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        mView = inflater.inflate(R.layout.new_order_details_fragment, null);
+        UserAuthenticationStateMachine mUserState = UserAuthenticationStateMachineProducer.getInstance(getActivity());
+        mSiteId = mUserState.getSiteId();
+        mTenantId = mUserState.getTenantId();
+        mIsEditMode = getArguments().getBoolean(EDIT_MODE);
+        ButterKnife.inject(this, mView);
+        NewOrderManager.count = 0;
+        return mView;
+    }
+
+    private Subscriber<Object> getEventSubscriber() {
+        return new Subscriber<Object>() {
+            @Override
+            public void onCompleted() {
+
+            }
+
+            @Override
+            public void onError(Throwable e) {
+
+            }
+
+            @Override
+            public void onNext(Object o) {
+                if (o instanceof Boolean) {
+                    mIsEditMode = (Boolean) o;
+                    updateEditMode(mIsEditMode);
+                }
+            }
+        };
+    }
+
+    private void updateEditMode(Boolean editMode) {
+        mProductsAdapter.setEditMode(editMode);
+        mProductsAdapter.notifyDataSetChanged();
+        mProductLookup.setVisibility(editMode ? View.VISIBLE : View.GONE);
+    }
+
+    private void setUpViews() {
+        mAdapter = new ProductSuggestionAdapter(getActivity());
+        mProductLookup.setAdapter(mAdapter);
+        mProductsAdapter = new NewOrderProductAdapter(this);
+        if (mOrder != null && mOrder.getItems() != null) {
+            mProductsAdapter.addData(mOrder);
+            updateEditMode(mIsEditMode);
+        }
+
+
+        mOrderProducts.setAdapter(mProductsAdapter);
+        mProductLookup.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence charSequence, int i, int i2, int i3) {
+
+            }
+
+            @Override
+            public void onTextChanged(final CharSequence charSequence, int start, int before, int count) {
+                if (mSubscription != null && !mSubscription.isUnsubscribed()) {
+                    mSubscription.unsubscribe();
+                }
+
+                mSubscription = AndroidObservable.bindFragment(NewOrderCreateFragment.this, NewOrderManager.getInstance().getProductSuggestion(charSequence.toString(), mTenantId, mSiteId))
+                        .subscribe(new Subscriber<ProductSearchResult>() {
+                            @Override
+                            public void onCompleted() {
+
+                            }
+
+                            @Override
+                            public void onError(Throwable e) {
+                                Toast.makeText(getActivity(), "Error Searching for product " + charSequence, Toast.LENGTH_SHORT).show();
+                            }
+
+                            @Override
+                            public void onNext(ProductSearchResult productSearchResult) {
+                                List<Product> data = productSearchResult.getItems();
+                                mAdapter.clear();
+                                mAdapter.addAll(data);
+                                mAdapter.notifyDataSetChanged();
+                            }
+                        });
+            }
+
+            @Override
+            public void afterTextChanged(Editable editable) {
+
+            }
+        });
+
+        mProductLookup.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
+                Product product = (Product) adapterView.getItemAtPosition(position);
+                OrderItem orderItem = new OrderItem();
+                orderItem.setProduct(convertProduct(product));
+                orderItem.setQuantity(1);
+                NewOrderItemEditFragment newOrderItemEditFragment = NewOrderItemEditFragment.getInstance(orderItem, mOrder.getId(), false);
+                newOrderItemEditFragment.setOnEditDoneListener(NewOrderCreateFragment.this);
+                newOrderItemEditFragment.show(getFragmentManager(), "");
+            }
+        });
+
+        mOrderProducts.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> adapterView, View view, int position, long l) {
+                if (mIsEditMode && adapterView.getItemAtPosition(position) instanceof OrderItemRow) {
+                    OrderItem orderItem = ((OrderItemRow) adapterView.getItemAtPosition(position)).orderItem;
+                    NewOrderItemEditFragment newOrderItemEditFragment = NewOrderItemEditFragment.getInstance(orderItem, mOrder.getId(), true);
+                    newOrderItemEditFragment.setOnEditDoneListener(NewOrderCreateFragment.this);
+                    newOrderItemEditFragment.show(getFragmentManager(), "");
+                }
+            }
+        });
     }
 
     @Override
