+++ MozuAndroidInStoreAssistant/app/src/main/java/com/mozu/mozuandroidinstoreassistant/app/utils/FulfillmentUtils.java
@@ -16,5 +16,4 @@ public class FulfillmentUtils {
     }
 
 
-
 }
