+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/new_order_product_item.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <com.mozu.mozuandroidinstoreassistant.app.layout.order.NewOrderItemLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:gravity="center_vertical"
+    android:layout_height="72dp"
     android:background="@android:color/white"
-    android:paddingLeft="24dp"
-    android:layout_height="72dp">
+    android:gravity="center_vertical"
+    android:orientation="vertical"
+    android:paddingLeft="24dp">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -14,57 +14,58 @@
 
         <TextView
             android:id="@+id/product_code"
-            android:layout_weight="1"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:layout_weight="1"/>
 
         <LinearLayout
-            android:layout_weight="3"
             android:layout_width="0dp"
-            android:orientation="vertical"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:layout_weight="3"
+            android:orientation="vertical">
 
             <TextView
                 android:id="@+id/product_name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"/>
-            <TextView
-                android:visibility="gone"
-                android:id="@+id/product_variation"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
 
             <TextView
-                android:visibility="gone"
-                android:layout_marginTop="8dp"
+                android:id="@+id/product_variation"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:visibility="gone"/>
+
+            <TextView
                 android:id="@+id/product_discount"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
+                android:layout_height="wrap_content"
+                android:layout_marginTop="8dp"
+                android:visibility="gone"/>
         </LinearLayout>
 
         <TextView
             android:id="@+id/fulfillment_type"
-            android:layout_weight="1"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:layout_weight="1"/>
 
         <TextView
             android:id="@+id/product_price"
-            android:layout_weight="1"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:layout_weight="1"/>
 
         <TextView
             android:id="@+id/product_quantity"
-            android:layout_weight="1"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:layout_weight="1"/>
 
         <LinearLayout
-            android:layout_weight="1"
             android:layout_width="0dp"
-            android:orientation="vertical"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:orientation="vertical">
 
             <TextView
                 android:id="@+id/product_total"
@@ -72,10 +73,10 @@
                 android:layout_height="wrap_content"/>
 
             <TextView
-                android:textColor="@color/mozu_color"
                 android:id="@+id/product_discount_price"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
+                android:layout_height="wrap_content"
+                android:textColor="@color/mozu_color"/>
         </LinearLayout>
     </LinearLayout>
