+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/new_order_shipping_item.xml
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <com.mozu.mozuandroidinstoreassistant.app.layout.order.NewOrderShippingItemLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:padding="24dp"
-    android:layout_height="wrap_content">
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    android:padding="24dp">
 
     <TextView
-        android:text="@string/shipping"
-        android:gravity="center_vertical"
         android:layout_width="wrap_content"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent"
+        android:gravity="center_vertical"
+        android:text="@string/shipping"/>
 
     <LinearLayout
-        android:layout_marginLeft="-3dp"
-        android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:layout_marginLeft="-3dp"
+        android:orientation="horizontal">
 
         <Spinner
             android:id="@+id/shipping_spinner"
@@ -24,10 +24,10 @@
             android:layout_height="match_parent"/>
 
         <ProgressBar
-            android:visibility="gone"
             android:id="@+id/shipping_progress"
             android:layout_width="wrap_content"
-            android:layout_height="match_parent"/>
+            android:layout_height="match_parent"
+            android:visibility="gone"/>
     </LinearLayout>
