+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/neworder_header_layout.xml
@@ -1,49 +1,49 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="horizontal"
               android:layout_width="match_parent"
-              android:gravity="center_vertical"
-              android:background="@android:color/white"
+              android:layout_height="64dp"
               android:layout_marginLeft="16dp"
-              android:paddingLeft="24dp"
-              android:layout_height="64dp">
+              android:background="@android:color/white"
+              android:gravity="center_vertical"
+              android:orientation="horizontal"
+              android:paddingLeft="24dp">
 
     <TextView
-        android:text="@string/product_code_header"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="@string/product_code_header"/>
 
     <TextView
-        android:text="@string/product_name_header"
         android:id="@+id/product_name"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
         android:layout_weight="3"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:text="@string/product_name_header"/>
 
     <TextView
-        android:text="@string/product_fullfilment_header"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="@string/product_fullfilment_header"/>
 
     <TextView
-        android:text="@string/product_price_header"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="@string/product_price_header"/>
 
     <TextView
-        android:text="@string/product_quantity_header"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="@string/product_quantity_header"/>
 
     <TextView
-        android:text="@string/product_total_header"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="@string/product_total_header"/>
 
 
 </LinearLayout>
