+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/neworder_total_item.xml
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:orientation="horizontal"
               android:layout_width="match_parent"
-              android:layout_height="match_parent">
+              android:layout_height="match_parent"
+              android:orientation="horizontal">
 
     <TextView
-        android:layout_weight="7"
         android:id="@+id/column_name"
-        android:gravity="right"
-        android:layout_marginRight="24dp"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_marginRight="24dp"
+        android:layout_weight="7"
+        android:gravity="right"/>
 
     <TextView
-        android:textColor="@color/mozu_color"
         android:id="@+id/column_value"
-        android:layout_weight="1"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"/>
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:textColor="@color/mozu_color"/>
 
 </LinearLayout>
