+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/order_detail_notes_fragment.xml
@@ -126,7 +126,6 @@
                 android:layout_marginEnd="8dp"
                 android:layout_marginRight="8dp"
                 android:background="@drawable/floating_action_button"
-                android:elevation="4dp"
                 android:text="+"
                 android:textSize="25sp"/>
