+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/orderfulfillment_dropdown_resource.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <TextView xmlns:android="http://schemas.android.com/apk/res/android"
-          android:textStyle="normal"
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
-          android:textSize="20sp"/>
+          android:textSize="20sp"
+          android:textStyle="normal"/>
