+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/orderfulfillment_spinner_item.xml
@@ -5,11 +5,11 @@
               android:layout_height="48dp">
 
     <TextView
-        android:textStyle="normal"
-        android:gravity="center"
         android:id="@+id/order_fulfillment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:textSize="20sp"/>
+        android:gravity="center"
+        android:textSize="20sp"
+        android:textStyle="normal"/>
 
 </LinearLayout>
