+++ MozuAndroidInStoreAssistant/app/src/main/res/layout/product_detail_inventory_fragment.xml
@@ -1,41 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-
+<LinearLayout
     android:id="@+id/mainlayout"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+
     android:layout_width="@dimen/inventory_dialog_width"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <LinearLayout
         android:id="@+id/dialog_header"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:visibility="gone"
-        android:orientation="vertical">
+        android:orientation="vertical"
+        android:visibility="gone">
 
         <RelativeLayout
             android:layout_width="match_parent"
-            android:layout_marginBottom="16dp"
-            android:layout_height="wrap_content">
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="16dp">
 
             <TextView
                 android:layout_width="wrap_content"
-                android:textSize="14sp"
-                android:layout_alignParentLeft="true"
-                android:textColor="@color/light_gray_text"
                 android:layout_height="wrap_content"
-                android:text="@string/inventory_tab_name" />
+                android:layout_alignParentLeft="true"
+                android:text="@string/inventory_tab_name"
+                android:textColor="@color/light_gray_text"
+                android:textSize="14sp"/>
 
 
             <ImageView
-                android:layout_alignParentRight="true"
-                android:padding="0dp"
-                android:layout_margin="0dp"
                 android:id="@+id/imageView_close"
                 android:layout_width="wrap_content"
                 android:layout_height="16dp"
+                android:layout_alignParentRight="true"
+                android:layout_margin="0dp"
                 android:clickable="true"
+                android:padding="0dp"
                 android:src="@drawable/icon_close" />
 
         </RelativeLayout>
@@ -45,88 +46,88 @@
             android:layout_width="wrap_content"
             android:layout_height="24dp"
             android:gravity="center_vertical"
-            android:textSize="16sp"
-            android:textColor="@color/mozu_color" />
+            android:textColor="@color/mozu_color"
+            android:textSize="16sp"/>
 
         <TextView
             android:id="@+id/productName"
             android:layout_width="match_parent"
-            android:singleLine="true"
-            android:gravity="center_vertical"
             android:layout_height="48dp"
             android:layout_marginTop="5dp"
+            android:gravity="center_vertical"
+            android:singleLine="true"
             android:textSize="34sp" />
     </LinearLayout>
 
     <LinearLayout
-        android:layout_marginTop="10dp"
         android:id="@+id/product_variation_layout"
         android:layout_width="match_parent"
-        android:orientation="vertical"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:orientation="vertical">
 
         <TextView
             android:id="@+id/product_option_label"
-            android:paddingLeft="5dp"
-            android:text="@string/inventory_store_label"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            android:paddingLeft="5dp"
+            android:text="@string/inventory_store_label"/>
 
         <Spinner
             android:id="@+id/product_variation_spinner"
-            android:layout_height="wrap_content"
-            android:layout_width="match_parent" />
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"/>
     </LinearLayout>
 
     <LinearLayout
         android:id="@+id/inventory_header"
-        android:orientation="horizontal"
         android:layout_width="match_parent"
+        android:layout_height="64dp"
         android:gravity="center_vertical"
-        android:layout_height="64dp">
+        android:orientation="horizontal">
 
         <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
+            android:text="@string/prod_detail_inventory_table_code_header"
             android:textColor="@color/light_gray_text"
-            android:textSize="14sp"
-            android:text="@string/prod_detail_inventory_table_code_header" />
+            android:textSize="14sp"/>
 
         <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:textColor="@color/light_gray_text"
-            android:textSize="14sp"
             android:layout_weight="2"
-            android:text="@string/prod_detail_inventory_table_name_header" />
+            android:text="@string/prod_detail_inventory_table_name_header"
+            android:textColor="@color/light_gray_text"
+            android:textSize="14sp"/>
 
         <TextView
-            android:textColor="@color/light_gray_text"
-            android:textSize="14sp"
             android:layout_width="0dp"
-            android:gravity="right"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:text="@string/prod_detail_inventory_table_available_header" />
+            android:gravity="right"
+            android:text="@string/prod_detail_inventory_table_available_header"
+            android:textColor="@color/light_gray_text"
+            android:textSize="14sp"/>
 
         <TextView
-            android:textColor="@color/light_gray_text"
-            android:textSize="14sp"
             android:layout_width="0dp"
-            android:gravity="right"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:text="@string/prod_detail_inventory_table_on_hand_header" />
+            android:gravity="right"
+            android:text="@string/prod_detail_inventory_table_on_hand_header"
+            android:textColor="@color/light_gray_text"
+            android:textSize="14sp"/>
 
         <TextView
-            android:textColor="@color/light_gray_text"
-            android:textSize="14sp"
             android:layout_width="0dp"
-            android:gravity="right"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:text="@string/prod_detail_inventory_table_on_reserve_header" />
+            android:gravity="right"
+            android:text="@string/prod_detail_inventory_table_on_reserve_header"
+            android:textColor="@color/light_gray_text"
+            android:textSize="14sp"/>
     </LinearLayout>
 
     <View
@@ -149,8 +150,8 @@
             android:id="@+id/inventory_list"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:dividerHeight="0dp"
-            android:divider="@android:color/transparent" />
+            android:divider="@android:color/transparent"
+            android:dividerHeight="0dp"/>
     </com.mozu.mozuandroidinstoreassistant.app.views.LoadingView>
